[REFACTOR(CardViewUtil)] 패키지이동 및 적용
